refactor AutoFormatPlugin
